pyproject.toml
[project]
name = "multiragchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "huggingface-hub>=1.5.0",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
